Break charting code's dependency on LoopKit
