+++ Common/Extensions/ChartPoint.swift
@@ -7,7 +7,6 @@
 //
 
 import Foundation
-import LoopKit
 import SwiftCharts
 
 
@@ -96,15 +95,3 @@ extension ChartPoint {
     }
 
 }
-
-
-extension ChartPoint: TimelineValue {
-    public var startDate: Date {
-        if let dateValue = x as? ChartAxisValueDate {
-            return dateValue.date
-        } else {
-            return Date.distantPast
-        }
-    }
-}
-
