+++ Loop/Extensions/ChartPoint+Loop.swift
@@ -39,3 +39,16 @@ extension ChartPoint {
     }
 }
 
+
+
+extension ChartPoint: TimelineValue {
+    public var startDate: Date {
+        if let dateValue = x as? ChartAxisValueDate {
+            return dateValue.date
+        } else {
+            return Date.distantPast
+        }
+    }
+}
+
+
